Huge ZenEngine refactor
There was a lot of bad design decisions made. Fixing it, refactoring everything that needed change. I will need to establish better engine design before going forward

+++ Development/Include/Game/Game.h
@@ -11,6 +11,6 @@ class Game : public IEngineApplication
 {
 public:
 	void StartUp() override;
-	void Update(const uint32 deltaTime) override;
+	void Update(const double deltaTime) override;
 	void ShutDown() override;
 };
